Move assignments into the corresponding wer files
